renaming test etl now that it is in its own directory
